feat: 实现相册和图片管理功能v1版本
refactor: 重构后端实体类和DTO，统一使用Integer作为ID类型
fix: 修复文件上传路径和删除逻辑
style: 优化前端相册和图片页面样式
docs: 更新数据库脚本和注释, 将主键id改为int类型,并且设置为自增

+++ sidifensen_blog_backend/src/main/java/com/sidifensen/utils/JwtUtils.java
@@ -7,7 +7,6 @@ import cn.hutool.jwt.signers.JWTSignerUtil;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Component;
 
-
 import java.util.Date;
 import java.util.Map;
 import java.util.concurrent.TimeUnit;
@@ -24,7 +23,7 @@ public class JwtUtils {
      * @param isRememberMe 是否记住我
      * @return
      */
-    public String createToken(Long id,boolean isRememberMe) {
+    public String createToken(Integer id,boolean isRememberMe) {
         String token = JWT.create()
                 // 设置过期时间 rememberMe为true时，token有效期为7天，否则为1天
                 .setExpiresAt(new Date(System.currentTimeMillis() + TimeUnit.DAYS.toMillis(isRememberMe? 7 : 1)))
